chore: move constants out of opts Signed-off-by: Pete Muir <pmuir@bleepbleep.org.uk>
